--- test/Core.Tests/Geofence/GeofenceServiceFixture.cs
@@ -1,10 +1,0 @@
-using Rocket.Surgery.Extensions.Testing.Fixtures;
-
-namespace Rocket.Surgery.Airframe.Core.Tests.Geofence;
-
-internal class GeofenceServiceFixture : ITestFixtureBuilder
-{
-    public static implicit operator GeofenceService(GeofenceServiceFixture fixture) => fixture.Build();
-
-    private GeofenceService Build() => new GeofenceService();
-}
