--- test/Core.Tests/Geofence/GeofenceServiceTests.cs
@@ -1,7 +1,0 @@
-namespace Rocket.Surgery.Airframe.Core.Tests.Geofence
-{
-    public class GeofenceServiceTests
-    {
-
-    }
-}
